+++ docs/CHANGELOG.rst
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_
 and this project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+Changed
+~~~~~~~
+- Switch to new pyarrow.fs api instead of deprecated pyarrow.hdfs
 
 [2.11.0] - 2020-12-11
 ---------------------